refactor: 가격 및 날짜 포맷팅 로직을 core.formatter 패키지로 분리 및 이동
