+++ app/src/main/java/woowacourse/shopping/core/formatter/DateFormatter.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.core.formatter
+
+import java.time.LocalDate
+
+fun LocalDate.toFormattedDateString(): String = "${year}년 ${monthValue}월 ${dayOfMonth}일"
