+++ app/src/main/java/woowacourse/shopping/ui/recommendation/RecommendationScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.sp
 import woowacourse.shopping.R
 import woowacourse.shopping.core.designsystem.component.ShoppingItem
 import woowacourse.shopping.core.designsystem.component.layout.CommonFrame
-import woowacourse.shopping.core.designsystem.component.toPriceString
+import woowacourse.shopping.core.formatter.toPriceString
 import woowacourse.shopping.ui.cart.component.CartBottomBar
 import woowacourse.shopping.ui.uimodel.ProductUiModel
